requirements.txt
django==4.1.2
djangorestframework==3.14.0
django-cors-headers
djongo==1.3.6
certifi